Refactor: Rename React{ ⇒ .}Fragment
