+++ core/src/main/scala/japgolly/scalajs/react/feature/ReactFragment.scala
@@ -1,6 +1,7 @@
-package japgolly.scalajs.react.vdom
+package japgolly.scalajs.react.feature
 
 import japgolly.scalajs.react.{Key, raw => Raw}
+import japgolly.scalajs.react.vdom._
 import scala.scalajs.js
 
 object ReactFragment {
