+++ doc/changelog/1.3.0.md
@@ -29,11 +29,16 @@
 
     * React Context API implemented via `React.createContext` and `React.Context`.
 
+* Deprecated `ReactFragment` and replaced with `React.Fragment`.
 
 
 ## Migration
 
 ```sh
+find . -name '*.scala' -type f -exec perl -pi -e '
+  s/ReactFragment/React.Fragment/g;
+  ' {} +
+
 find . -path '*/test/scala/*.scala' -type f -exec perl -pi -e '
   s/getDOMNode.asElement/getDOMNode.asMounted().asElement()/g;
   s/getDOMNode.asText/getDOMNode.asMounted().asText()/g;
